feat: db modal split (#14470)
* refactor: split db modal file (#14436)

* split db modal file

* fix tests

* fix extra options

* update database id methodology
